running total updates not working

+++ Ledger/CoreData/Sortable.swift
@@ -16,7 +16,7 @@ protocol Sortable {
         get
         set
     }
-    var sortOrder: Double {
+    var sortOrder: Int {
         get
         set
     }
